requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
celery>=5.3
redis>=5.0
gunicorn>=21.2
python-dotenv>=1.0
Pillow>=10.0
requests>=2.31
google-generativeai>=0.3
sentence-transformers>=2.2
torch>=2.0 --index-url https://download.pytorch.org/whl/cpu
open_clip_torch>=2.20
nudenet>=3.2
numpy<2.0
onnxruntime>=1.16
django-redis>=5.3.0
